+++ src/index.ts
@@ -2,19 +2,21 @@ import alphabeticalOrderRule from './css-rules/alphabetical-order/index.js';
 import concentricOrderRule from './css-rules/concentric-order/index.js';
 import customOrderRule from './css-rules/custom-order/rule-definition.js';
 import noEmptyStyleBlocksRule from './css-rules/no-empty-blocks/rule-definition.js';
+import noTrailingZeroRule from './css-rules/no-trailing-zero/rule-definition.js';
 import noUnknownUnitRule from './css-rules/no-unknown-unit/rule-definition.js';
 import noZeroUnitRule from './css-rules/no-zero-unit/rule-definition.js';
 
 const vanillaExtract = {
   meta: {
     name: '@antebudimir/eslint-plugin-vanilla-extract',
-    version: '1.11.1',
+    version: '1.12.0',
   },
   rules: {
     'alphabetical-order': alphabeticalOrderRule,
     'concentric-order': concentricOrderRule,
     'custom-order': customOrderRule,
     'no-empty-style-blocks': noEmptyStyleBlocksRule,
+    'no-trailing-zero': noTrailingZeroRule,
     'no-unknown-unit': noUnknownUnitRule,
     'no-zero-unit': noZeroUnitRule,
   },
@@ -29,6 +31,7 @@ Object.assign(vanillaExtract.configs, {
     rules: {
       'vanilla-extract/concentric-order': 'error',
       'vanilla-extract/no-empty-style-blocks': 'error',
+      'vanilla-extract/no-trailing-zero': 'error',
       'vanilla-extract/no-unknown-unit': 'error',
       'vanilla-extract/no-zero-unit': 'error',
     },